# Improve performance of make_symbol_table