This is a continuation of #104. The `Interpreter` is constructed many times when fitting with lmfit and this calls `make_symbol_table`. This PR improves performance by taking many checks on the valid symbols in `math`, `sympy` and `builtins` out of the function.

Master:
```
In [2]: %timeit make_symbol_table()
172 µs ± 3.16 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
This PR:
```
In [59]: %timeit make_symbol_table()
11.3 µs ± 421 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Note: the (slighty) reduced coverage is in a part that was only modified by autopep and branches where numpy is not installed, but it seems numpy is installed when running coverage. If it is required to fix this, please suggest how to approach this.