@newville 2022-07-07
@eendebakpt thanks, looks very interesting. I am traveling and may not be able to review this sufficiently for a few weeks. I am not ignoring,  just may not get too quickly.  Feel free to ping me again on this if I'm going too slow. Thanks again.