@newville 2022-09-01
@eendebakpt Thanks again, this looks like an impressive speedup.  I am not actually concerned about the `coverage` report -- that's informative but does not really signal an error, especially with massive changes to code blocks. 

I verified locally on one machine that all `asteval` tests and all `lmfit` tests pass with this change.   

So, I will very gratefully merge this: thanks, awesome!